fix: add additional description for joins and value example
