+++ scripts/generate_python_relational_docs_examples.py
@@ -1,3 +1,25 @@
+DEFAULT_EXAMPLE = '''```python
+import duckdb
+
+duckdb_conn = duckdb.connect()
+
+rel = duckdb_conn.sql("""
+        select 
+            gen_random_uuid() as id, 
+            concat('value is ', case when mod(range,2)=0 then 'even' else 'uneven' end) as description,
+            range as value, 
+            now() + concat(range,' ', 'minutes')::interval as created_timestamp
+        from range(1, 10)
+    """
+)
+
+{code_example}
+```
+'''
+
+PLACEHOLDER_EXAMPLE = "```python\n{code_example}\n```"
+PLACEHOLDER_RESULT = "```{result_type}\n{result}\n```"
+
 CREATION_MEMBER_CODE_EXAMPLE_MAP = {
     'from_arrow': {
         'example': """
@@ -247,8 +269,6 @@ import duckdb
 
 duckdb_conn = duckdb.connect()
 
-duckdb_conn.sql("drop table if exists code_example")
-
 duckdb_conn.sql("create table code_example as select * from range(1,2) tbl(id)")
 
 rel = duckdb_conn.table("code_example")
@@ -297,7 +317,6 @@ import duckdb
 
 duckdb_conn = duckdb.connect()
 
-duckdb_conn.sql("drop table if exists code_example")
 duckdb_conn.sql("create table code_example as select * from range(1,2) tbl(id)")
 
 rel = duckdb_conn.view("code_example")
@@ -494,8 +513,6 @@ from uuid import uuid4
 
 duckdb_conn = duckdb.connect()
 
-duckdb_conn.sql("drop table if exists code_example")
-
 duckdb_conn.sql("""
         select 
             gen_random_uuid() as id, 
@@ -529,40 +546,48 @@ rel.filter("value = 10")
 """,
         'default': False,
     },
-    #     'insert_into': {
-    #         'example': '''
-    # import duckdb
-    #
-    # from datetime import datetime
-    # from uuid import uuid4
-    #
-    # duckdb_conn = duckdb.connect()
-    #
-    # duckdb_conn.sql("drop table if exists code_example")
-    #
-    # duckdb_conn.sql("""
-    #         select
-    #             gen_random_uuid() as id,
-    #             concat('value is ', case when mod(range,2)=0 then 'even' else 'uneven' end) as description,
-    #             range as value,
-    #             now() + concat(range,' ', 'minutes')::interval as created_timestamp
-    #         from range(1, 10)
-    #     """
-    # ).to_table("code_example")
-    #
-    # FIX ME
-    # rel = duckdb_conn.values(
-    #         uuid4(),
-    #         'value is even',
-    #         10,
-    #         datetime.now()
-    # )
-    #
-    # rel.filter("value = 10")
-    # ''',
-    #         'result': '',
-    #         'default': False,
-    #     },
+    'insert_into': {
+        'example': '''
+import duckdb
+
+from datetime import datetime
+from uuid import uuid4
+
+duckdb_conn = duckdb.connect()
+
+duckdb_conn.sql("""
+        select
+            gen_random_uuid() as id,
+            concat('value is ', case when mod(range,2)=0 then 'even' else 'uneven' end) as description,
+            range as value,
+            now() + concat(range,' ', 'minutes')::interval as created_timestamp
+        from range(1, 10)
+    """
+).to_table("code_example")
+
+rel = duckdb_conn.values(
+    [
+        uuid4(),
+        'value is even',
+        10,
+        datetime.now()
+    ]
+)
+
+rel.insert_into("code_example")
+
+duckdb_conn.table("code_example").filter("value = 10")
+    ''',
+        'result': """
+┌──────────────────────────────────────┬───────────────┬───────┬───────────────────────────────┐
+│                  id                  │  description  │ value │       created_timestamp       │
+│                 uuid                 │    varchar    │ int64 │   timestamp with time zone    │
+├──────────────────────────────────────┼───────────────┼───────┼───────────────────────────────┤
+│ 271c5ddd-c1d5-4638-b5a0-d8c7dc9e8220 │ value is even │    10 │ 2025-04-10 14:29:18.616379+02 │
+└──────────────────────────────────────┴───────────────┴───────┴───────────────────────────────┘
+""",
+        'default': False,
+    },
     'intersect': {
         'example': 'rel.intersect(other_rel=rel.set_alias("other_rel"))',
         'result': """
@@ -597,6 +622,64 @@ rel.count("*")
 └──────────────┘
 """,
         'default': True,
+        'additional_description': """
+
+Depending on how the `condition` parameter is provided, the JOIN clause generated is:
+- `USING`
+
+    ```python
+    import duckdb
+    
+    duckdb_conn = duckdb.connect()
+    
+    rel1 = duckdb_conn.sql("select range as id, concat('dummy 1', range) as text from range(1,10)")
+    rel2 = duckdb_conn.sql("select range as id, concat('dummy 2', range) as text from range(5,7)")
+    
+    rel1.join(rel2, condition="id", how="inner").sql_query()
+    ```
+    with following SQL:
+    ```sql
+    SELECT * 
+    FROM (
+            SELECT "range" AS id, 
+                concat('dummy 1', "range") AS "text" 
+            FROM "range"(1, 10)
+        ) AS unnamed_relation_41bc15e744037078 
+    INNER JOIN (
+            SELECT "range" AS id, 
+            concat('dummy 2', "range") AS "text" 
+            FROM "range"(5, 7)
+        ) AS unnamed_relation_307e245965aa2c2b 
+    USING (id)
+    ```
+- `ON`
+
+    ```python
+    import duckdb
+    
+    duckdb_conn = duckdb.connect()
+    
+    rel1 = duckdb_conn.sql("select range as id, concat('dummy 1', range) as text from range(1,10)")
+    rel2 = duckdb_conn.sql("select range as id, concat('dummy 2', range) as text from range(5,7)")
+    
+    rel1.join(rel2, condition=f"{rel1.alias}.id = {rel2.alias}.id", how="inner").sql_query()
+    ```
+    with the following SQL:
+    ```sql
+    SELECT * 
+    FROM (
+            SELECT "range" AS id, 
+                concat('dummy 1', "range") AS "text" 
+            FROM "range"(1, 10)
+        ) AS unnamed_relation_41bc15e744037078 
+    INNER JOIN (
+            SELECT "range" AS id, 
+            concat('dummy 2', "range") AS "text" 
+            FROM "range"(5, 7)
+        ) AS unnamed_relation_307e245965aa2c2b 
+    ON ((unnamed_relation_41bc15e744037078.id = unnamed_relation_307e245965aa2c2b.id))
+    ```
+""",
     },
     'limit': {
         'example': 'rel.limit(1)',
@@ -1111,4 +1194,12 @@ created_timestamp: [[2025-04-10 09:54:24.015000Z,2025-04-10 09:55:24.015000Z,202
         'result': 'The data is exported to a Parquet file, named code_example.parquet',
         'default': True,
     },
+}
+
+CODE_EXAMPLE_MAP = {
+    **CREATION_MEMBER_CODE_EXAMPLE_MAP,
+    **DEFINITION_MEMBER_CODE_EXAMPLE_MAP,
+    **TRANSFORMATION_MEMBER_CODE_EXAMPLE_MAP,
+    **FUNCTION_MEMBER_CODE_EXAMPLE_MAP,
+    **OUTPUT_MEMBER_CODE_EXAMPLE_MAP,
 }
