+++ scripts/generate_python_relational_docs_methods.py
@@ -1,0 +1,142 @@
+DEFINITION_MEMBER_LIST = [
+    "columns",
+    "describe",
+    "description",
+    "dtypes",
+    "explain",
+    "query",
+    "set_alias",
+    "alias",
+    "shape",
+    "show",
+    "sql_query",
+    "type",
+    "types",
+]
+TRANSFORMATION_MEMBER_LIST = [
+    "aggregate",
+    "apply",
+    "cross",
+    "except_",
+    "filter",
+    "insert",
+    "insert_into",
+    "intersect",
+    "join",
+    "limit",
+    "map",
+    "order",
+    "project",
+    "select",
+    "sort",
+    "union",
+    "update",
+]
+FUNCTION_MEMBER_LIST = [
+    "any_value",
+    "arg_max",
+    "arg_min",
+    "avg",
+    "bit_and",
+    "bit_or",
+    "bit_xor",
+    "bitstring_agg",
+    "bool_and",
+    "bool_or",
+    "count",
+    "cume_dist",
+    "dense_rank",
+    "distinct",
+    "favg",
+    "first",
+    "first_value",
+    "fsum",
+    "geomean",
+    "histogram",
+    "lag",
+    "last",
+    "last_value",
+    "lead",
+    "list",
+    "max",
+    "mean",
+    "median",
+    "min",
+    "mode",
+    "n_tile",
+    "nth_value",
+    "percent_rank",
+    "product",
+    "quantile",
+    "quantile_cont",
+    "quantile_disc",
+    "rank",
+    "rank_dense",
+    "row_number",
+    "select_dtypes",
+    "select_types",
+    "std",
+    "stddev",
+    "stddev_pop",
+    "stddev_samp",
+    "string_agg",
+    "sum",
+    "unique",
+    "value_counts",
+    "var",
+    "var_pop",
+    "var_samp",
+    "variance",
+]
+OUTPUT_MEMBER_LIST = [
+    "arrow",
+    "close",
+    "create",
+    "create_view",
+    "df",
+    "execute",
+    "fetch_arrow_reader",
+    "fetch_arrow_table",
+    "fetch_df_chunk",
+    "fetchall",
+    "fetchdf",
+    "fetchmany",
+    "fetchnumpy",
+    "fetchone",
+    "pl",
+    "record_batch",
+    "tf",
+    "to_arrow_table",
+    "to_csv",
+    "to_df",
+    "to_parquet",
+    "to_table",
+    "to_view",
+    "torch",
+    "write_csv",
+    "write_parquet",
+]
+
+CREATION_MEMBER_LIST = [
+    'from_arrow',
+    'from_csv_auto',
+    'from_df',
+    'from_parquet',
+    'from_query',
+    'query',
+    'read_csv',
+    'read_json',
+    'read_parquet',
+    'sql',
+    'table',
+    'table_function',
+    'values',
+    'view',
+]
+
+PY_RELATION_MEMBERS = (
+    DEFINITION_MEMBER_LIST
+    + TRANSFORMATION_MEMBER_LIST
+    + FUNCTION_MEMBER_LIST
+    + OUTPUT_MEMBER_LIST
+)
